feat post 패키지명 오타 리팩토링 jap -> jpa
