+++ src/main/java/com/leesang/mylocaldiary/post/jpa/entity/Photo.java
@@ -1,32 +1,22 @@
-package com.leesang.mylocaldiary.post.jap.entity;
+package com.leesang.mylocaldiary.post.jpa.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
-
-import java.math.BigDecimal;
 
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Builder
-public class Place {
+public class Photo {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private String name;
-
-    @Column(precision = 10, scale = 7)
-    private BigDecimal latitude;
-
-    @Column(precision = 10, scale = 7)
-    private BigDecimal longitude;
+    private String imageUrl;
 
     private int orders;
-
-    private String thumbnailImage;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "post_id")
