+++ src/main/java/com/leesang/mylocaldiary/post/jpa/repository/PlaceRepository.java
@@ -1,6 +1,6 @@
-package com.leesang.mylocaldiary.post.jap.repository;
+package com.leesang.mylocaldiary.post.jpa.repository;
 
-import com.leesang.mylocaldiary.post.jap.entity.Place;
+import com.leesang.mylocaldiary.post.jpa.entity.Place;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PlaceRepository extends JpaRepository<Place, Long> {
