+++ src/main/java/com/leesang/mylocaldiary/post/jpa/repository/PostRepository.java
@@ -1,6 +1,6 @@
-package com.leesang.mylocaldiary.post.jap.repository;
+package com.leesang.mylocaldiary.post.jpa.repository;
 
-import com.leesang.mylocaldiary.post.jap.entity.Post;
+import com.leesang.mylocaldiary.post.jpa.entity.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
